app/data_product_framework-master/python/requirements/bazel.txt
pip==9.0.3
pylint==1.8.4
Pillow==5.1.0
numpy==1.14.3
scipy==1.1.0
opencv-python==3.4.0.12
tensorflow-gpu==1.8
